Added composite parser for AWD 1&2. Details below:
The AWDParser wraps an instance of either AWD1Parser or AWD2Parser,
depending on which claims the data (checking AWD2Parser first.) This
way, we avoid cases where either one of AWD1Parser and AWD2Parser
claims an *.awd file based solely on it's suffix, and then ends up
not being able to read the actual data.

AWDParser will claim .awd files, and files that have data that
conforms to either AWD1Parser or AWD2Parser, and when parsing starts
pick the one that will work, delegating all methods and events
between that concrete parser and the environment.