Switched AssetLoader/ResourceLoadSession. Details:
AssetLoader is promoted as a "user-space" class, which end-users are
encouraged to use and should as such be as easy to use as possible and
provide many useful features. It could however only load single files
before this change.

ResourceLoadSession was regarded an internal class of the asset lib and
not promoted as a user class, but was actually more useful since it could
load dependencies.

This commit switches the places (and roles) of these two classes. The
ResourceLoadSession becomes AssetLoader (with retained functionality)
while the old AssetLoader becomes a new internal class called
"SingleResourceLoader".

A misc package is created for SingleResourceLoader and ResourceDependency,
and since this leaves AssetLibrary the only class in the library package,
it's moved to the loading base package (and the now empty library package
is removed.)

This leaves three classes directly in the loading package, all of which are
promoted as user space classes (Loader3D, AssetLoader now with automatic
dependency loading, and AssetLibrary.)

Classes that basic users should not have to interface with are in sub
packages of the loading package, just like it should be! :)