Added AssetLoaderToken, details below:
Loading classes (starting with AssetLoader and above in the stack) will now
return an instance of the class AssetLoaderToken. This token can be used as
an event dispatcher on which to set listeners for loader events for that
specific loader.

When using the AssetLoader directly this makes little sense. When using the
Loader3D or AssetLibrary where several loads can be started at the same time,
the AssetLoaderToken provides a way to only listen for events on a specific
load (instead of listening on AssetLibrary or the Loader3D instance.)

The benefit of returning this token rather than returning the actual
AssetLoader instance is that the token is a completely read-only black box
with which the user cannot accidentally disrupt the load process.

EVENT BUFFERING
Since by nature the token will only be available to the user after the load
has been initialized, there will be cases where the load will have finished
before the user can set any listeners. An example is a small embedded file
where the parser will have completed synchronously before the parseData()
call returns.

To solve this, the AssetLoaderToken will buffer events for which there are
no listeners, and dispatch them all in order after an event listener has
been added. If none is added, these events will never be dispatched by the
token, which will eventually be garbage collected.