Brought back thourough check to loads with no suffix
For some reason, there was an if statement effectively disabling the thorough
file type check for resources that do not have a file suffix. Refactored
SingleFileLoader slightly so that files are always loaded regardless of whether
getParserFromSuffix() returns anything, and that getParserFromData() is always
used in parse() when no parser has been found in a previuos step.