Removed hard-coded stuff in AWD2Parser, details:
AWD2Parser previously always created SkinnedSubGeometry instances, even if the
mesh was not skinned. This has now changed, and regular SubGeometry instances
will be created unless there is both a joint weight and joint index data stream
in the file.

Furthermore, the number of joints per vertex is no longer hard-coded as 3, but
instead defined as numWeights/numVerts.

git-svn-id: https://away3d.googlecode.com/svn/trunk/broomstick/Away3D@3279 6da1ee2c-932f-0410-8bb4-73b2b089c45a