Improved dependency handling, details below:
Parsers can now pause themselves and wait for dependencies to be retrieved,
using the ParserBase method pauseAndRetrieveDependencies(). This will stop
executing proceedParsing() and dispatch the READY_FOR_DEPENDENCIES event on
which the AssetLoader will start loading dependencies. When all dependencies
have been recursively retrieved, it will resume parsing using the ParserBase
resumeParsingAfterDependencies() method.

This commit adds this new system, as well as an implementation of it in the
AWDParser/AWD2Parser.