Added raw-data dependency loading, details:
Parsers can now use the regular dependency-queueing mechanism to load raw-data
dependencies, meaning that the file that is loaded will not go through a parser
but instead be returned to the requesting parser directly after the load.

The raw data (ByteArray) will be given to the parent resolveDependency() method
as part of the ResourceDependency instance, accessible via the data property.

This can be used for example by the OBJParser to load MTL files for instance,
without the need for a "MTLParser".