Fixed null reference issues in AssetLoader
It would previously check for URL-mapped data in the loader context, without
first making sure that there was indeed a context object. It would also assume
that dependency objects always had a request object, which is not always the
case (i.e. when using the parseData() method for embedded data.)