OBJParser now uses dependency system for MTL file
This was previuosly not really possible, but since the dependency system's
recent revamp, the MTL file can be loaded as "raw data" and hence does not
require a separate parser. Also, it can be loaded without having to fake that
the parsing was complete, whereas at the time the OBJParser was created, all
dependency loading was delayed until after the parsing was done.