Implemented resolution of asset name colisions:
AssetLibrary will now check for collisions when adding assets and when assets
are renamed, and if collisions exist, will ask it's "naming strategy" class to
resolve them.

This commit includes only a NumSuffixNamingStrategy class which adds a numeric
suffix to names to make sure there are no collisions. More strategies to come.