Optimized naming strategies, added create(), details:
To optimize the naming strategies, they are now allowed to have internal state,
which required them to be created as a new instance for every library, instead
of a shared static instance. For this, a create() method has been added that
will serve as a self-factory method.

The NumSuffixNamingStrategy now stores the next available suffix per basename,
meaning that little (or no) looping is required every time a collision happens,
which proved to be absolutely necessary for performance in cases where the same
name occurs very many times (e.g. unnamed skeleton poses.)