Revert "Mesh's addSubMesh() method is now protected instead of private."
This reverts commit 845984c08d50a1809e565c940e2103e0c602a40b.

Comments by David Lenaerts:
SubMeshes are a completely managed resource, and you should never be able to
create or destroy them manually, as the only thing that will happen is that
there'll be a misalignment between the Mesh and the Geometry.

Worse, it introduces orphan SubGeometry instances that can't be properly
disposed of. You're supposed to call geometry.addSubGeometry, and the SubMesh
will be created automatically for you.