Resource limit errors
Created InvalidTextureError and implemented it in Texture3DProxy.as and
BackgroundImageRenderer.as
The idea is to centralize verbose and inform the user of runtime errors
caused by Stage3D limitations wherever possible. So far the engine
throws too little errors. It must inform users about such limitations
using custom errors.