Did a straight conversion of OBJParser to the new texture system. Details:
The OBJ parser has a lot of legacy code, and uses a very "tangly" way of dealing
with dependencies. The entire parser will probably need to be rebuilt.
Disregarding that, this update will try to make sure that the OBJParser works
with the new texture system, and any further updates are postponed.