Improved the way alphaPremultiplied and blendMode works together.
Previously setting blendMode after alphaPremultiplied would overwrite the blend
factors, effectively disabling the impact of the alphaPremultiplied property,
and vice versa. This fix adds an updateBlendFactors() function that is called
whenever alphaPremultiplied or blendMode is invoked. It is called instantly,
instead of invalidating the values, since the function is so cheap and contains
code that will throw an error on incorrect blend modes, which is something that
should happen when the property is set.