Fixed issue with primitive validation. Fixes #257
Previously the Geometry class was never accessed in the render loop, since
the sub-geometries are accessed via their sub-meshes instead. This meant that
the PrimitiveBase (geometry sub-class) was never given a chance to validate
it's geometry.

This fix introduces a validate() method on Geometry which is invoked by the
subMeshes getter on Mesh. This makes sure that primitive is given a chance
to validate and potentially change the number of sub-geometries, and in turn
the sub-meshes, before they are returned.

This means that changing properties on primitives now correctly triggers a
geometry update at the next render.