Fixed issue with calculating bounds for some Geometry instances. Fixes #273
Since the calculation in fromGeometry() used +/- Infinity as starting points
and an if/else assuming that the same vertex could not define opposite bounds
(e.g. left AND right) it would break when the first vertex was indeed an
extreme. This fix starts with assuming the first vertex defines the extremes
on all axes and then uses that when comparing the other vertex positions.