Fixes incorrect reading of length fields. Important! Will break files.
This fix corrects a mistake in the AWD2 parser that was unfortunately also made
in several generators (anything using libawd.) This is why the issue was not
discovered earlier. The error was that attribute length fields were read as
uint16 fields, instead of uint32 per the specification. Leaving this would have
meant that property values cannot be longer than 65535 bytes, which can cause
problems in some cases.

If you're using the 3ds Max AWD exporter, version 1.0.2 generates correct files
that will work with this version of the parser.