Added SubGeometry.autoGenerateDummyUVs flag, to toggle automatic UV generation
In some cases, a model does not come with UV coordinates, resulting in a crash
when texture materials are used. This is currently circumvented by some parsers
that create dummy UV buffer content. However, since this happens in the parser,
there is no way of knowing whether it's actually even needed (as in the case of
only using color materials.) This changeset implements the same type of feature
on the SubGeometry instead, meaning that a dummy UV buffer can be automatically
generated on-demand instead of up-front, similar to how auto-generation of
normals and tangents works.