Fixed crash issue on first backbuffer config in AIR. Closes #230
In AIR particularly, the Context3D seems to be returned fast enough that
View3D.render() will not yet have been invoked. Since that's the function that
usually triggers the first backbuffer configuration, no width and height (or
value for depthAndStencil) will have been set when the context is returned,
resulting in an invalid call to configureBackBuffer (with invalid values for
the width, height and depthAndStencil arguments.)

This fix checks the width and height before calling configureBackBuffer(),
thereby making sure that it does not happen prematurely.