Fixed issue with validation in Object3D.
Previously the pos, rot and scale dirty flags would be reset in the transform
getter, even though the actual validation happens in updateTransform(). This
mean that if a sub-class was to invoke updateTransform() directly (which makes
sense in some cases) the position, rotation and scale vectors were not marked
as being valid again. This in turn would lead to notifyPositionChange() (and
the equivalents for rotation and scale) to bail early since position, rotation
and scale were already marked as dirty, preventing any more events from being
dispatched.