Changed order of transform and component invalidation in component setters.
Previously, changing position, rotation or scale properties would first dispatch
an event, and then set the dirty flags to true. This meant that if an event
listener reacted to the event by requesting a re-validation (e.g. by querying
the transform or sceneTransform properties) the validation would have already
happened by the time the dirty flags were set to true. This in turn mean that
next time the position, rotation or scale was changed the notify function would
bail early (dirty==true) and not dispatch anymore events.

Hence, any system depending on these events to re-validate the transform would
only execute once, since next update would not result in an event being
dispatched.

This simple fix changes the order such that the dirty flags will first be set
to true, and the event dispatched second. This way, an listeners that trigger
an update of the transform matrix will properly result in the object being
marked as clean again.