Fixed issue in how Loader3D handled LOAD_ERROR events.
Previously listeners would be prematurely removed (even though user correctly
handled the LOAD_ERROR event) meaning that if the load was allowed to continue
no more events would be caught and able to bubble up to the user.

This fix correctly waits with removing the event listeners until it's known
whether it's possible to continue loading, and also implements the correct
behavior for LoaderEvent.preventDefault().