Now properly cleaning up listeners on LOAD_ERROR
Previously Loader3D would not clean up error listeners after a LOAD_ERROR
event unless isDefaultPrevented() was true. This is fine for dependency
errors, but if the error event was caused by being unable to load the base
file, the load operation would stop here even if default behavior was not
prevented, and hence the event listeners should be properly cleaned up.