Implemented functioning error handling in loaders.
This new error handling system functions much better than before although it is
a little bit arcane in comparison. It doesn't use the regular event dispatcher
system but instead has a custom callback system for error events. This lets
AssetLoader know whether client code is listening for it's events by checking
hasEventListener() (without risking the listener being AssetLibraryBundle or
Loader3D) and can hance decide whether to throw an error more reliably.

Second, the callbacks return whether or not the events were handled as a result
of those callbacks, i.e. if there were listeners on the objects that own the
callbacks (Loader3D and/or AssetLibraryBundle) and thus AssetLoader will know
after the callbacks return whether the event was handled at all.