Solved issue with specular method in OBJParser. Fixes #64
The default behavior of DefaultScreenPass (via DefaultMaterialBase) is to
overwrite the properties of a new specular method with those of the existing
one. This is so that setting the specularMethod (or any other method) property
does not change the settings, but only the method by which those settings are
used. This is great on most cases, but OBJParser uses a SpecularMethod instance
as a sort of value object, and needs it's properties to be kept intact as it is
assigned to the material.

This fix prevents the pass from overwriting the new method's settings by making
sure that the previous shading method is null, meaning there are no properties
that can be copied over.