Removed dispose() override in TextureMaterial. Fixes #316
TextureMaterial.dispose() was previously calling dispose on the screen pass,
and then invoking the super method. However, since the screen pass is in the
material's passes list, the super class would try to dispose it again, causing
a null reference error. Because the super class is capable of disposing the
pass itself, this override is not necessary.