Removed DATA_COMPLETE in favor of DEPENDENCY_COMPLETE
The DATA_COMPLETE was only ever used internally between SingleFileLoader and
AssetLoader, and was a risk of confusion for users. It was also misleading
since it was in fact dispatched after parsing was done, not just when the data
had loaded. Since we already have DEPENDENCY_COMPLETE, and that's exactly what
it means when SingleFileLoader finishes loading and parsing something, it made
sense to reuse that event in this case too.