Delayed dependency resolve() after successful loading.
There are several problems with resolving dependencies straight after they
finished loading. Some formats create additional dependencies as part of their
dependency resolution (e.g. OBJParser resolving the MTL file) and others that
dispatch asset events will have those supressed if the dependency being
resolved was configured to suppress events (because as far as the AssetLoader
knows, that is still the loading dependency.)

This fix delays the actual resolve() call until the next retrieveNext() call,
by doing it as a result of the stack popping.