Fixed normals/tangents generation in cylinder. Closes #209
Now properly respects the slope of the lateral surface,
meaning that cones will now get slanting normals instead
of normals that are parallel with the base plane. Also
fixes issue when radius = 0, which it will always be for
one segment in cones. This previously resulted in zero
normals, which is no longer the case.