Add coordinateSystem property to PerspectiveLens.as
Away3D appears to default to a left-handed coordinate system, but other systems (e.g. Alternativa3D, OpenSceneGraph, etc) defaults to a right-handed projection. For compatibility with such systems I have added an option to PerspectiveLens which allows the handedness of the coordinate system to be switched.

The default state of the projection will not be affected by this change.